requirements-celery.txt
celery==5.3.6 
redis==5.0.1 
flask-celery-helper==1.1.0
